Using newDirectoryStream and Consumer
An example of using the static method "Files.newDirectoryStream" to iterate over the files in that particular directory. The filenames found are printed to the console using the functional interface "Consumer". We don't have to use the interface Consumer, I've done it this way so as to illustrate how to use the Consumer interface (which I always forget).