Use Files.mismatch to compare two files
Uses the static method Files.mismatch to compare two files. The method finds and returns the position of the first mismatched byte in the content of two files, or the value -1L if there is no mismatch (the two files are the same). The position indicated will be in the inclusive range of 0L up to the size (in bytes) of the smaller file. This is a quick and easy way of comparing two files.